Algo- Sort  - Bubble Sort  - Insertion Sort  - Selection Sort
